+++ lighthouse/pipeline/descriptor.py
@@ -1,6 +1,8 @@
 import yaml
 import os
 
+from lighthouse.pipeline.helper import remove_args_and_opts, update_filename
+
 
 class PipelineDescriptor:
     """
@@ -8,9 +10,14 @@ class PipelineDescriptor:
     This class is responsible for parsing the pipeline description from a YAML file,
     and keeping a list of stages for comsumption by the Driver.
 
+    The format here is just text. The main job of this class is to handle includes,
+    to verify that the files for the stages exist, normalize their paths, etc.
+    The actual validation of the stages is left to the Driver and the stages themselves.
+
     Format is:
     Pipeline:
       - pass: PassName
+      - transform: TransformFile.py[gen=generator_name,seq=sequence_name]{opt1=val1 opt2=val2}
       - transform: TransformFile.mlir
       - include: OtherPipeline.yaml
       - bundle: BundleName
@@ -29,8 +36,26 @@ class PipelineDescriptor:
             )
 
     def _normalize_include_path(self, filename) -> str:
-        """Uses the path of the includer to determine the path of the included."""
-        return os.path.normpath(os.path.join(os.path.dirname(self.filename), filename))
+        """
+        Finds the file in some standard locations, in order:
+            * The path of the descriptor file that includes it. This allows for relative includes.
+            * The path of the Lighthouse schedule module, where all the standard pipelines are located.
+        """
+        filename = remove_args_and_opts(filename)
+        descriptor_path = os.path.normpath(os.path.dirname(self.filename))
+        schedule_module_path = os.path.normpath(
+            os.path.join(os.path.dirname(__file__), "../schedule")
+        )
+
+        file = os.path.join(descriptor_path, filename)
+        if not os.path.exists(file):
+            file = os.path.join(schedule_module_path, filename)
+            if not os.path.exists(file):
+                raise ValueError(
+                    f"Included pipeline descriptor file does not exist: {filename} \
+                        (searched in {descriptor_path} and {schedule_module_path})"
+                )
+        return file
 
     def _parse_stages(self) -> None:
         """
@@ -50,11 +75,11 @@ class PipelineDescriptor:
             elif "transform" in stage:
                 # Transforms need to be MLIR files, and need to exist.
                 filename = self._normalize_include_path(stage["transform"])
-                if not os.path.exists(filename):
-                    raise ValueError(f"Transform file does not exist: {filename}")
-                elif not filename.endswith(".mlir"):
-                    raise ValueError(f"Transform file must be an MLIR file: {filename}")
-                self.stages.append(filename)
+                if not filename.endswith(".mlir") and not filename.endswith(".py"):
+                    raise ValueError(
+                        f"Transform file must be an MLIR or Python file: {filename}"
+                    )
+                self.stages.append(update_filename(stage["transform"], filename))
 
             elif "pass" in stage:
                 # Passes are just strings, let the pass manager validate.
@@ -74,13 +99,8 @@ class PipelineDescriptor:
     def _include_pipeline(self, filename: str) -> None:
         """
         Helper function to include another pipeline descriptor file.
-        Include path is RELATIVE to the file including.
         """
         filename = self._normalize_include_path(filename)
-        if not os.path.exists(filename):
-            raise ValueError(
-                f"Included pipeline descriptor file does not exist: {filename}"
-            )
         included_pipeline = PipelineDescriptor(filename)
         self.stages.extend(included_pipeline.get_stages())
